Add support to load TTF TrueType fonts using FreeType
Add support to load TTF TrueType fonts using FreeType.

* **NengenFileUtil.java**
  - Add import statements for FreeType classes.
  - Add a new method `loadTTF` to load TTF fonts using FreeType.

* **GameFont.java**
  - Add a constructor to accept FreeType font data.
  - Add a method to load FreeType font data.
  - Update `delete` method to handle null texture.

* **CharacterData.java**
  - Add a constructor to accept FreeType character data.

* **pom.xml**
  - Add dependency for `lwjgl-freetype` version `3.3.6`.

+++ src/main/java/visuals/rendering/text/GameFont.java
@@ -4,16 +4,32 @@ import visuals.lwjgl.render.Texture;
 
 public final class GameFont {
 
+	public static long FONT_INDEX = 0;
+
+	/**
+	 * The global index of the font.
+	 */
+	private final long fontIndex;
+
 	private final String name;
 	private final int fontSize;
-	private final CharacterData[] characterDatas;
+	private CharacterData[] characterDatas;
 	private final Texture texture;
 
-	public GameFont(String name, int fontSize, Texture texture) {
+	public GameFont(String name, int fontSize, Texture texture, long fontIndex) {
 		this.name = name;
 		this.fontSize = fontSize;
 		characterDatas = new CharacterData[128];
 		this.texture = texture;
+		this.fontIndex = fontIndex;
+	}
+
+	public GameFont(String name, int fontSize, CharacterData[] characterDatas, long fontIndex) {
+		this.name = name;
+		this.fontSize = fontSize;
+		this.characterDatas = characterDatas;
+		this.texture = null;
+		this.fontIndex = fontIndex;
 	}
 
 	public String getName() {
@@ -28,12 +44,22 @@ public final class GameFont {
 		return characterDatas;
 	}
 
+	public void characterDatas(CharacterData[] characterDatas) {
+		this.characterDatas = characterDatas;
+	}
+
 	public Texture texture() {
 		return texture;
 	}
 
 	public void delete() {
-		texture.delete();
+		if (texture != null) {
+			texture.delete();
+		}
+	}
+
+	public static long getFontIndex() {
+		return FONT_INDEX++;
 	}
 
 }
